+++ Source/Steam Library Manager/Definitions/Updater.cs
@@ -11,7 +11,7 @@ namespace Steam_Library_Manager.Definitions
         public static string LatestVersionDownloadURL = "https://raw.githubusercontent.com/RevoLand/Steam-Library-Manager/master/Binaries/Steam%20Library%20Manager.exe";
 
         // Current version, should be increased with each release
-        public static Version CurrentVersion = new Version(System.Diagnostics.FileVersionInfo.GetVersionInfo(System.Reflection.Assembly.GetExecutingAssembly().Location).ProductVersion);
+        public static Version CurrentVersion = new Version(System.Windows.Forms.Application.ProductVersion);
 
         // Latest SLM version, will be updated from UpdateLink
         public static Version LatestVersion = new Version();