+++ Source/Steam Library Manager/Functions/SLM.cs
@@ -1,5 +1,4 @@
-﻿using Microsoft.Win32;
-using System;
+﻿using System;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
@@ -96,61 +95,23 @@ namespace Steam_Library_Manager.Functions
 
         public static void LoadSteam()
         {
-            Steam.UpdateSteamInstallationPathAsync();
-            Steam.PopulateLibraryCMenuItems();
-            Steam.PopulateAppCMenuItems();
+            try
+            {
+                Steam.UpdateSteamInstallationPathAsync();
+                Steam.PopulateLibraryCMenuItems();
+                Steam.PopulateAppCMenuItems();
 
-            Steam.Library.GenerateLibraryList();
+                Steam.Library.GenerateLibraryList();
+            }
+            catch (Exception ex)
+            {
+                Logger.LogToFile(Logger.LogType.Library, ex.ToString());
+            }
         }
 
         public static void OnClosing()
         {
             Settings.SaveSettings();
-        }
-
-        public static string Get45PlusFromRegistry()
-        {
-            const string subkey = @"SOFTWARE\Microsoft\NET Framework Setup\NDP\v4\Full\";
-
-            using (RegistryKey ndpKey = RegistryKey.OpenBaseKey(RegistryHive.LocalMachine, RegistryView.Registry32).OpenSubKey(subkey))
-            {
-                if (ndpKey != null && ndpKey.GetValue("Release") != null)
-                {
-                    return CheckFor45PlusVersion((int)ndpKey.GetValue("Release"));
-                }
-            }
-
-            return "Unknown";
-        }
-
-        static string CheckFor45PlusVersion(int releaseKey)
-        {
-            if (releaseKey >= 460798)
-                return "4.7 or later";
-            if (releaseKey >= 394802)
-                return "4.6.2";
-            if (releaseKey >= 394254)
-            {
-                return "4.6.1";
-            }
-            if (releaseKey >= 393295)
-            {
-                return "4.6";
-            }
-            if ((releaseKey >= 379893))
-            {
-                return "4.5.2";
-            }
-            if ((releaseKey >= 378675))
-            {
-                return "4.5.1";
-            }
-            if ((releaseKey >= 378389))
-            {
-                return "4.5";
-            }
-
-            return "No 4.5 or later version detected";
         }
 
         public class Library
@@ -162,6 +123,7 @@ namespace Steam_Library_Manager.Functions
                     // If we have a backup library(s)
                     if (Properties.Settings.Default.backupDirectories == null)
                         return;
+
                     if (Properties.Settings.Default.backupDirectories.Count == 0)
                         return;